simplify managing differences of default nginx config location across platforms
