--- include_darwin.go
@@ -1,8 +1,0 @@
-// +build darwin
-
-package nginxconf
-
-func init() {
-	// https://www.cyberciti.biz/faq/freebsd-install-nginx-webserver/
-	nginxConfPrefix = "/usr/local/etc/nginx"
-}
