--- include_freebsd.go
@@ -1,8 +1,0 @@
-// +build freebsd
-
-package nginxconf
-
-func init() {
-	// https://www.cyberciti.biz/faq/freebsd-install-nginx-webserver/
-	nginxConfPrefix = "/usr/local/etc/nginx"
-}
