--- include_netbsd.go
@@ -1,8 +1,0 @@
-// +build netbsd
-
-package nginxconf
-
-func init() {
-	// https://www.netbsd.mx/nginx-php.html
-	nginxConfPrefix = "/usr/pkg/etc/nginx"
-}
