--- include_nix.go
@@ -1,7 +1,0 @@
-// +build !windows !freebsd !netbsd !solaris !illumos
-
-package nginxconf
-
-func init() {
-	nginxConfPrefix = "/etc/nginx"
-}
